Moved SubscriberDirectoryProcessor to JsonProcessor
