+++ src/Platform.Eda.Cli/Commands/ConfigureEda/ConfigureEdaCommand.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using CaptainHook.Domain.Results;
 using McMaster.Extensions.CommandLineUtils;
 using Microsoft.Rest;
+using Platform.Eda.Cli.Commands.ConfigureEda.JsonProcessor;
 using Platform.Eda.Cli.Commands.ConfigureEda.Models;
 
 namespace Platform.Eda.Cli.Commands.ConfigureEda
