+++ src/Platform.Eda.Cli/Commands/ConfigureEda/JsonProcessor/ISubscribersDirectoryProcessor.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using CaptainHook.Domain.Results;
 using Platform.Eda.Cli.Commands.ConfigureEda.Models;
 
-namespace Platform.Eda.Cli.Commands.ConfigureEda
+namespace Platform.Eda.Cli.Commands.ConfigureEda.JsonProcessor
 {
     public interface ISubscribersDirectoryProcessor
     {
