+++ src/Tests/Platform.Eda.Cli.Tests/Commands/ConfigureEda/SubscribersDirectoryProcessorTests.cs
@@ -6,9 +6,8 @@ using CaptainHook.Api.Client.Models;
 using CaptainHook.Domain.Results;
 using Eshopworld.Tests.Core;
 using FluentAssertions;
-using Moq;
 using Newtonsoft.Json.Linq;
-using Platform.Eda.Cli.Commands.ConfigureEda;
+using Platform.Eda.Cli.Commands.ConfigureEda.JsonProcessor;
 using Platform.Eda.Cli.Commands.ConfigureEda.Models;
 using Platform.Eda.Cli.Common;
 using Xunit;
